+++ shanoir-ng-datasets/src/main/java/org/shanoir/ng/solr/model/ShanoirSolrDocument.java
@@ -88,15 +88,15 @@ public class ShanoirSolrDocument {
 	
 	@Field
 	@Indexed
-	private Float sliceThickness;
+	private Double sliceThickness;
 	
 	@Field
 	@Indexed
-	private Float pixelBandwidth;
+	private Double pixelBandwidth;
 	
 	@Field
 	@Indexed
-	private Float magneticFieldStrength;
+	private Double magneticFieldStrength;
 	
 	@Field
 	@Indexed(name="tags", type="string")
@@ -108,8 +108,8 @@ public class ShanoirSolrDocument {
 	
 	public ShanoirSolrDocument (String id, Long datasetId, String datasetName, String datasetType, String datasetNature,
 			Date datasetCreationDate, String examinationComment, Date examinationDate,
-			String subjectName, String studyName, Long studyId, String centerName, Float sliceThickness,
-			Float pixelBandwidth, Float magneticFieldStrength) {
+			String subjectName, String studyName, Long studyId, String centerName, Double sliceThickness,
+			Double pixelBandwidth, Double magneticFieldStrength) {
 		this.id = id;
 		this.datasetId = datasetId;
 		this.datasetName = datasetName;
@@ -283,27 +283,27 @@ public class ShanoirSolrDocument {
 		this.centerName = centerName;
 	}
 
-	public Float getSliceThickness() {
+	public Double getSliceThickness() {
 		return sliceThickness;
 	}
 
-	public void setSliceThickness(Float sliceThickness) {
+	public void setSliceThickness(Double sliceThickness) {
 		this.sliceThickness = sliceThickness;
 	}
 
-	public Float getPixelBandwidth() {
+	public Double getPixelBandwidth() {
 		return pixelBandwidth;
 	}
 
-	public void setPixelBandwidth(Float pixelBandwidth) {
+	public void setPixelBandwidth(Double pixelBandwidth) {
 		this.pixelBandwidth = pixelBandwidth;
 	}
 
-	public Float getMagneticFieldStrength() {
+	public Double getMagneticFieldStrength() {
 		return magneticFieldStrength;
 	}
 
-	public void setMagneticFieldStrength(Float magneticFieldStrength) {
+	public void setMagneticFieldStrength(Double magneticFieldStrength) {
 		this.magneticFieldStrength = magneticFieldStrength;
 	}
